working after major refactor
