+++ a_front_end/index.html
@@ -29,7 +29,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 <!-- jQuery 1.6.0 -->
 
 <!-- local version for offline testing -->
-<script type="text/javascript" src="2_front_end/jquery.min.js"></script>
+<script type="text/javascript" src="a_front_end/jquery.min.js"></script>
 
 <!-- online version hosted by Google -->
 <!--
@@ -37,7 +37,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  -->
 
 
-<script type="text/javascript" src="2_front_end/jquery.corner.js"></script>
+<script type="text/javascript" src="a_front_end/jquery.corner.js"></script>
 <script type="text/javascript">
 $(document).ready(function() {
   // for rounded corners
@@ -46,7 +46,7 @@ $(document).ready(function() {
 </script>
 
 
-<link rel="stylesheet" href="2_front_end/edu-python-title.css"/>
+<link rel="stylesheet" href="a_front_end/edu-python-title.css"/>
 
 </head>
 
@@ -86,9 +86,9 @@ z = ["Alice", "Bob", "Charlie"]
 aliasing relationships:</p>
 
 <center>
-<img src="2_front_end/alias-screenshot.png"/>
+<img src="a_front_end/alias-screenshot.png"/>
 
-<h3><a href="2_front_end/tutor.html">Go play with the Online Python Tutor!</a></h3>
+<h3><a href="a_front_end/tutor.html">Go play with the Online Python Tutor!</a></h3>
 </center>
 
 </div>
@@ -107,7 +107,7 @@ executing on a series of graded test inputs, and showing what tests
 passed and failed.</p>
 
 <center>
-<img src="2_front_end/grading-375.png"/>
+<img src="a_front_end/grading-375.png"/>
 </center>
 
 <p>The above screenshot shows passed and failed tests.  The user can
@@ -118,9 +118,9 @@ the program failed on a particular test.</p>
 
 <ul>
 
-<li><a href="2_front_end/question.html?two-sum">Two-sum</a></li>
-<li><a href="2_front_end/question.html?reverse">Reverse list</a></li>
-<li><a href="2_front_end/question.html?remove-dups">Remove duplicate characters</a></li>
+<li><a href="a_front_end/question.html?two-sum">Two-sum</a></li>
+<li><a href="a_front_end/question.html?reverse">Reverse list</a></li>
+<li><a href="a_front_end/question.html?remove-dups">Remove duplicate characters</a></li>
 
 </ul>
 
@@ -142,9 +142,9 @@ some sample debugging problems:</p>
 
 <ul>
 
-<li><a href="2_front_end/question.html?debug-ireverse">In-place reverse</a></li>
-<li><a href="2_front_end/question.html?debug-bsearch">Binary search</a></li>
-<li><a href="2_front_end/question.html?debug-mergesort">Mergesort</a></li>
+<li><a href="a_front_end/question.html?debug-ireverse">In-place reverse</a></li>
+<li><a href="a_front_end/question.html?debug-bsearch">Binary search</a></li>
+<li><a href="a_front_end/question.html?debug-mergesort">Mergesort</a></li>
 
 </ul>
 
@@ -164,9 +164,9 @@ code."</i> Here are some sample optimization problems:</p>
 
 <ul>
 
-<li><a href="2_front_end/question.html?optimize-sum">Greatest sum</a></li>
-<li><a href="2_front_end/question.html?optimize-find-dups">Find duplicates</a></li>
-<li><a href="2_front_end/question.html?optimize-search">List search</a></li>
+<li><a href="a_front_end/question.html?optimize-sum">Greatest sum</a></li>
+<li><a href="a_front_end/question.html?optimize-find-dups">Find duplicates</a></li>
+<li><a href="a_front_end/question.html?optimize-search">List search</a></li>
 
 </ul>
 
@@ -179,8 +179,8 @@ code."</i> Here are some sample optimization problems:</p>
 
 <p>Teachers can easily create new practice problems by writing them in a
 lightweight plain text format.  For example, <a
-href="2_front_end/questions/reverse.txt">reverse.txt</a> provides the specification
-for the <a href="2_front_end/question.html?reverse">Reverse list</a> problem.</p>
+href="a_front_end/questions/reverse.txt">reverse.txt</a> provides the specification
+for the <a href="a_front_end/question.html?reverse">Reverse list</a> problem.</p>
 
 <p>The problem specification format allows constraints like <i>"code
 diffs must be less than N lines"</i> (used for debugging problems) and
