+++ a_front_end/tutor.html
@@ -111,20 +111,20 @@ Write your Python code here:
 <p style="margin-top: 25px;">
 Then try some programming questions:<br/>
 Solve:
-<a href="2_front_end/question.html?two-sum">Two-sum</a> |
-<a href="2_front_end/question.html?reverse">Reverse list</a> |
-<a href="2_front_end/question.html?remove-dups">Remove duplicate chars</a>
+<a href="a_front_end/question.html?two-sum">Two-sum</a> |
+<a href="a_front_end/question.html?reverse">Reverse list</a> |
+<a href="a_front_end/question.html?remove-dups">Remove duplicate chars</a>
 
 <br/>Debug:
-<a href="2_front_end/question.html?debug-ireverse">In-place reverse</a> |
-<a href="2_front_end/question.html?debug-bsearch">Binary search</a> |
-<a href="2_front_end/question.html?debug-mergesort">Mergesort</a>
+<a href="a_front_end/question.html?debug-ireverse">In-place reverse</a> |
+<a href="a_front_end/question.html?debug-bsearch">Binary search</a> |
+<a href="a_front_end/question.html?debug-mergesort">Mergesort</a>
 
 <br/>Optimize:
 
-<a href="2_front_end/question.html?optimize-sum">Greatest sum</a> |
-<a href="2_front_end/question.html?optimize-find-dups">Find duplicates</a> |
-<a href="2_front_end/question.html?optimize-search">List search</a>
+<a href="a_front_end/question.html?optimize-sum">Greatest sum</a> |
+<a href="a_front_end/question.html?optimize-find-dups">Find duplicates</a> |
+<a href="a_front_end/question.html?optimize-search">List search</a>
 
 </p>
